GET/PATCH User status

+++ controllers/feed.js
@@ -2,6 +2,7 @@ const fs = require('fs');
 const path = require('path');
 const { validationResult } = require('express-validator');
 const Post = require('../models/post');
+const User = require('../models/user');
 
 exports.getPosts = (req, res, next) => {
 	// pagination
@@ -54,19 +55,29 @@ exports.createPost = (req, res, next) => {
 	const imageUrl = req.file.path;
 	const title = req.body.title;
 	const content = req.body.content;
+	let creator;
+
 	const post = new Post({
 		title: title,
 		content: content,
 		imageUrl: imageUrl,
-		creator: { name: 'LibertySky' },
+		creator: req.userId,
 	});
 	post
 		.save()
-		.then((result) => {
-			// console.log(result);
+		.then(() => {
+			return User.findById(req.userId);
+		})
+		.then((user) => {
+			creator = user;
+			user.posts.push(post);
+			return user.save();
+		})
+		.then(() => {
 			res.status(201).json({
 				message: 'Post created successfully!',
-				post: result,
+				post: post,
+				creator: { _id: creator._id, name: creator.name },
 			});
 		})
 		.catch((err) => {
@@ -124,6 +135,13 @@ exports.editPost = (req, res, next) => {
 				error.statusCode = 404;
 				throw error;
 			}
+
+			if (post.creator.toString() !== req.userId) {
+				const error = new Error('Not authorized');
+				error.statusCode = 403;
+				throw error;
+			}
+
 			if (imageUrl !== post.imageUrl) {
 				clearImage(post.imageUrl);
 			}
@@ -154,11 +172,25 @@ exports.deletePost = (req, res, next) => {
 				error.statusCode = 404;
 				throw error;
 			}
+
+			if (post.creator.toString() !== req.userId) {
+				const error = new Error('Not authorized');
+				error.statusCode = 403;
+				throw error;
+			}
+
 			clearImage(post.imageUrl);
 			return Post.findByIdAndRemove(postId);
 		})
-		.then((result) => {
-			res.status(200).json({ message: 'Post deleted.', post: result });
+		.then(() => {
+			return User.findById(req.userId);
+		})
+		.then((user) => {
+			user.posts.pull(postId);
+			return user.save();
+		})
+		.then(() => {
+			res.status(200).json({ message: 'Post deleted.' });
 		})
 		.catch((err) => {
 			if (!err.statusCode) {
